chore(aft): Stop embedding `pub`
The Dart team recommends against embedding pub tooling and since they inline AWS+Google secrets in their source code, it can trip up GitHub when uploading aft binaries.
