+++ packages/aft/lib/aft.dart
@@ -12,7 +12,6 @@ export 'src/commands/exec_command.dart';
 export 'src/commands/generate/generate_command.dart';
 export 'src/commands/link_command.dart';
 export 'src/commands/list_packages_command.dart';
-export 'src/commands/pub_command.dart';
 export 'src/commands/publish_command.dart';
 export 'src/commands/run_command.dart';
 export 'src/commands/version_bump_command.dart';
