+++ packages/aft/pubspec.yaml
@@ -27,10 +27,6 @@ dependencies:
   meta: ^1.7.0
   mustache_template: ^2.0.0
   path: any
-  pub:
-    git:
-      url: https://github.com/dart-lang/pub.git
-      ref: 6cbbec2abbf6a54074ae1005c06a26dfb14a86c8
   pub_semver: ^2.1.1
   pubspec_parse: ^1.2.0
   smithy: ">=0.4.0+4 <0.5.0"
